chore: move IntoBytes into utils file
